tgbot/requirements.txt
python-telegram-bot>=21.0
aiohttp>=3.9.0
httpx>=0.27.0
